metrics are now in a special folder so that registry can work better
